+++ Pandora/Paradigm/Structure/Specific/Stack.hs
@@ -1,12 +1,14 @@
+{-# OPTIONS_GHC -fno-warn-orphans #-}
+
 module Pandora.Paradigm.Structure.Specific.Stack (Stack, push, top, pop, filter, linearize) where
 
-import Pandora.Core.Functor (type (:.), type (:=))
+import Pandora.Core.Functor (Variant (Co))
 import Pandora.Core.Transformation (type (~>))
 import Pandora.Paradigm.Basis.Maybe (Maybe (Just, Nothing))
 import Pandora.Paradigm.Basis.Predicate (Predicate (Predicate))
 import Pandora.Paradigm.Basis.Twister (Twister (Twister), untwist)
 import Pandora.Paradigm.Inventory.State (fold)
-import Pandora.Paradigm.Controlflow.Joint.Interpreted (Interpreted (Primary, unwrap))
+import Pandora.Paradigm.Controlflow.Joint.Schemes.UT (UT (UT))
 import Pandora.Pattern.Category ((.))
 import Pandora.Pattern.Functor.Covariant (Covariant ((<$>), (<$$>)))
 import Pandora.Pattern.Functor.Alternative (Alternative ((<+>)))
@@ -20,43 +22,39 @@ import Pandora.Pattern.Functor.Divariant (($))
 import Pandora.Pattern.Object.Setoid ((?))
 
 -- | Linear data structure that serves as a collection of elements
-newtype Stack a = Stack (Maybe :. Twister Maybe := a)
+type Stack = UT 'Co 'Co (Twister Maybe) Maybe
 
-instance Covariant Stack where
-	f <$> Stack stack = Stack $ f <$$> stack
+instance Covariant (UT 'Co 'Co (Twister Maybe) Maybe) where
+	f <$> UT stack = UT $ f <$$> stack
 
-instance Pointable Stack where
-	point = Stack . Just . point
+instance Pointable (UT 'Co 'Co (Twister Maybe) Maybe) where
+	point = UT . Just . point
 
-instance Alternative Stack where
-	Stack x <+> Stack y = Stack $ x <+> y
+instance Alternative (UT 'Co 'Co (Twister Maybe) Maybe) where
+	UT x <+> UT y = UT $ x <+> y
 
-instance Avoidable Stack where
-	empty = Stack Nothing
+instance Avoidable (UT 'Co 'Co (Twister Maybe) Maybe) where
+	empty = UT Nothing
 
-instance Applicative Stack where
-	Stack f <*> Stack x = Stack $ f <**> x
+instance Applicative (UT 'Co 'Co (Twister Maybe) Maybe) where
+	UT f <*> UT x = UT $ f <**> x
 
-instance Traversable Stack where
-	Stack stack ->> f = Stack <$> stack ->>> f
-
-instance Interpreted Stack where
-	type Primary Stack a = Maybe :. Twister Maybe := a
-	unwrap (Stack stack) = stack
+instance Traversable (UT 'Co 'Co (Twister Maybe) Maybe) where
+	UT stack ->> f = UT <$> stack ->>> f
 
 push :: a -> Stack a -> Stack a
-push x (Stack stack) = Stack $ (Twister x . Just <$> stack) <+> (point . point) x
+push x (UT stack) = UT $ (Twister x . Just <$> stack) <+> (point . point) x
 
 top :: Stack ~> Maybe
-top (Stack stack) = extract <$> stack
+top (UT stack) = extract <$> stack
 
 pop :: Stack ~> Stack
-pop (Stack stack) = Stack $ stack >>= untwist
+pop (UT stack) = UT $ stack >>= untwist
 
 filter :: Predicate a -> Stack a -> Stack a
-filter (Predicate p) = Stack . fold empty
+filter (Predicate p) = UT . fold empty
 	(\now new -> p now ? Just (Twister now new) $ new)
 
 -- | Transform any traversable structure into a stack
 linearize :: Traversable t => t ~> Stack
-linearize = Stack . fold Nothing (\x -> Just . Twister x)
+linearize = UT . fold Nothing (\x -> Just . Twister x)